backend/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
beautifulsoup4==4.12.2
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
multidict==6.0.4
orjson==3.8.12
psutil==5.9.5
pycodestyle==2.10.0
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.22.0
yarl==1.9.2
